clox: tweak emoji use so they line up better
iTerm's behavior with emoji is totally inscrutable to me, sometimes
displaying them as 2 columns and some as one. This means that the colons
often didn't line up, because the flag emojis are regional indicators
but the globe was a single-codepoint emoji, and text is hard. Fix this
by just using the UN flag for UTC, so at least they're all flags.